+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## master (unreleased)
 
-* No entry
+* Create Session objects on each connections (history)
 
 ## v1.4.0 (2017-11-24)